# lib/modules: Short-circuit unmatchedDefns when configs is empty