@roberth 2021-11-03
I've spotted some low hanging fruit that I've added here because it's simple enough. Will make separate PRs for other improvements though, because ofborg stats aren't comparable as it merges with (mutable) master before each run.