@roberth 2021-11-03
@Infinisil  That's all, as far as I'm concerned.

There's potential for more laziness, allowing `mkOption` to be replaced by bottom, by removing the `count` call and making `unmatchedDefns` tree-shaped, but I'm not going there with this PR. It seems rather too invasive because that data structure also affects types.