@andir 2021-12-11
I haven't invested much time but apparently this broke the way `extraArgs` are passed to module. My `morph`  deployments now don't get the `name` attribute passed to modules anymore and thus I am unable to extract the current nodes name in a module (without resorting to other module system parameters).  It looks like @eadwu ran into the same issue but simply reverted it locally?