@andir 2021-12-11
> @andir Are you sure it isn't fixed by https://github.com/NixOS/nixpkgs/pull/148315?
> 
> I don't know about eadwu's https://github.com/NixOS/nixpkgs/pull/144022#ref-commit-b2c6bf5, but it doesn't seem to be part of any ref anymore.

I think you are right. It must be that other PR. I somehow didn't find that.